requirements.txt
fastapi
uvicorn
python-dotenv
mistralai
faiss-cpu
pydantic
pytest
pydantic-settings